beagle_x15: Fix /dev/block/ file names
Partition files naming scheme was changed in kernel v5.0 due to moving
of some device tree definitions to this file:

    arch/arm/boot/dts/dra7-l4.dtsi

Because of this, we will have different file paths in AOSP for getting
partitions "by-name". Let's reflect it in fstab definitions, so that we
can support kernel v5.0+ (in future releases or when user provides
custom kernel with version 5.0 or above in KERNELDIR).

Change-Id: I052775068e1294f3087d4bae84919d5821db072e
Signed-off-by: Sam Protsenko <semen.protsenko@linaro.org>