beagle_x15: Fix boot_fit.img generation
Although bootfitimage module is added in PRODUCT_PACKAGES, it's not
being built on "make" command, and subsequent "make bootfitimage"
invocation is necessary. Let's make bootfitimage a prerequisite of
droidcore rule instead of referencing it in PRODUCT_PACKAGES. This way
boot_fit.img is guaranteed to be built on "make" command.

This is pretty much the same as it's done for HiKey board here:

    device/linaro/hikey/build/tasks/dtimage.mk

but in our case we add phony module name to droidcore prerequisites, not
file path. This way we can still run 'make bootfitimage' separately if
we want to.

Change-Id: I09fa9854437f64d35400311eadb6cdaec46fb144
Signed-off-by: Sam Protsenko <semen.protsenko@linaro.org>