Include dtb into boot.img
Android Boot Image v2 allows us to include our dtb files in it. For that
to happen we need to provide the directory with our *.dtb files, and
those will be squashed into single dtb.img file by Android build system
and added to "DTB" area of boot.img (so that DT blobs are concatenated
one after the other). Other way to do so would be to provide previously
generated DTBO image. Let's stick to the first way for now.

Change-Id: I3b6d969c0fa7e921acb9bb441e5a1861b408d1dd
Signed-off-by: Sam Protsenko <semen.protsenko@linaro.org>