Add board-info.txt file
This file is further copied as "android-info.txt" to
$ANDROID_PRODUCT_OUT and used by "fastboot flashall" to check if
"fastboot getvar product" returns board name from the list in this file.
Corresponding fix [1] is added to U-Boot bootloader as well. The same
values [2] were added to board-info.txt for consistency, as all those
boards are used with beagle_x15 AOSP target.

This way CheckRequirements() in FlashAll:Flash() works as expected,
checking if we are trying to flash images to the board that images were
built for.

No need to add that "board-info.txt" file to TARGET_BOARD_INFO_FILE, it
will be used by build system by default.

[1] https://patchwork.ozlabs.org/patch/1137022/
[2] https://gitlab.denx.de/u-boot/u-boot/blob/v2019.07/board/ti/am57xx/board.c#L538

Change-Id: Ib90f3ed118488cd6a0f9aae8150b1965f6c1b2cd
Signed-off-by: Sam Protsenko <semen.protsenko@linaro.org>