Enable AVB
Enable Android Verified Boot 2.0 for vbmeta.img to be generated. We need
to specify boot.img precise size so that AVB meta info can be added in
the footer. See [1,2] for details.

This change is needed for fastbootd to function properly, as it can't
flash partitions when androidboot.verifiedbootstate=orange is not set
(see [3]), which is done by 'avb verify' command in U-Boot (see [4]).
And 'avb verify' requires vbmeta.img to be flashed to vbmeta partition.

[1] https://android.googlesource.com/platform/external/avb/+/master/README.md
[2] https://gitlab.denx.de/u-boot/u-boot/blob/v2019.10-rc1/doc/android/avb2.txt
[3] https://android.googlesource.com/platform/system/core/+/refs/tags/android-q-preview-5/fastboot/device/utility.cpp#200
[4] https://gitlab.denx.de/u-boot/u-boot/blob/v2019.10-rc1/common/avb_verify.c#L149

Change-Id: I977fba70ad25955c9f4785d5f766c0362df083fa
Signed-off-by: Sam Protsenko <semen.protsenko@linaro.org>